automatically backed up by learn

+++ lib/create.sql
@@ -8,14 +8,14 @@ CREATE TABLE project (
   id INTEGER PRIMARY KEY,
   title TEXT,
   category TEXT,
-  funding_goal INTEGER,
+  funding_goal REAL,
   start_date DATE,
   end_date DATE
 )
 
 CREATE TABLE pledge (
   id INTEGER PRIMARY KEY,
-  amount FLOAT,
+  amount REAL,
   user_id INTEGER,
   project_id INTEGER
 )
